requirements.txt
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0
google-generativeai==0.7.2
python-dotenv==1.0.0
